refactor(scripts): split icons script into separate files, add new package.json#scripts entries
