+++ scripts/icons/optimize.ts
@@ -1,0 +1,71 @@
+/**
+ * Optimizes SVGs by running them through SVGO + Iconify cleanup.
+ */
+
+import { readFile, readdir, writeFile } from 'node:fs/promises'
+import { resolve } from 'node:path'
+import { flavorEntries } from '@catppuccin/palette'
+import { SVG, cleanupSVG, runSVGO } from '@iconify/tools'
+import { consola } from 'consola'
+
+try {
+  consola.info('Optimizing SVG files...')
+  Promise.all(flavorEntries.map(async ([flavor]) => {
+    const flavorPath = resolve('icons', flavor)
+    const svgs = await readdir(flavorPath)
+    await Promise.all(svgs.map(async (s) => {
+      const svgPath = resolve(flavorPath, s)
+      const str = await readFile(svgPath, 'utf8')
+      const svg = new SVG(str)
+      cleanupSVG(svg)
+      runSVGO(svg, { plugins: [
+        'cleanupAttrs',
+        'cleanupEnableBackground',
+        'cleanupIds',
+        'cleanupListOfValues',
+        'cleanupNumericValues',
+        'collapseGroups',
+        'convertColors',
+        'convertEllipseToCircle',
+        'convertShapeToPath',
+        'convertStyleToAttrs',
+        'convertTransform',
+        'inlineStyles',
+        'mergePaths',
+        'mergeStyles',
+        'minifyStyles',
+        'moveElemsAttrsToGroup',
+        'removeComments',
+        'removeDesc',
+        'removeDoctype',
+        'removeEditorsNSData',
+        'removeEmptyAttrs',
+        'removeEmptyContainers',
+        'removeEmptyText',
+        'removeHiddenElems',
+        'removeMetadata',
+        'removeNonInheritableGroupAttrs',
+        'removeRasterImages',
+        'removeScriptElement',
+        'removeStyleElement',
+        'removeTitle',
+        'removeUnknownsAndDefaults',
+        'removeUnusedNS',
+        'removeUselessDefs',
+        'removeUselessStrokeAndFill',
+        'removeXMLProcInst',
+        'sortAttrs',
+        'sortDefsChildren',
+      ], multipass: true })
+      await writeFile(
+        svgPath,
+        svg.toPrettyString()
+          .replaceAll(/#[A-F0-9]{6}/gi, s => s.toLowerCase()),
+      )
+    }))
+  }))
+  consola.success('Optimized SVG files.')
+}
+catch (error) {
+  consola.error('Optimization failed: ', error)
+}
